Changes to Player Super Class
Added a few Leveling Methods and also GiveAmmo, GiveExp, and GiveHealth
functions. To make compatible I also had to change the GetComponent in
ZombieController to look for <Player> and not <PlayerController> like we
had before. What's an easy fix to this, or do we have to rewrite a few
lines?