Leaning Mechanic (WIP)
Date Modified: 5/5/2017 @ 11:28 pm

Changelog: Added the beginnings of the PlayerLean() function. Was
working fine before, need to work out the glitches now. Unity says that
you can't have multiple transforms in the Debug Console, or something of
the sort when I run the program and press 'q' or 'e' (Leaning Mechanic).